% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{generateHeatmapMeanPowBand}
\alias{generateHeatmapMeanPowBand}
\title{compute the mean power analysis over the frequency band using the multitaper results for a condition
for a list of analysis windows}
\usage{
generateHeatmapMeanPowBand(
  epochList,
  multitaperResults,
  windowParams,
  analysisWindows,
  seizureCondition
)
}
\arguments{
\item{epochList}{list of Matrix or Epoch object. iEEG data matrix or Epoch object.}

\item{multitaperResults}{(numeric)  A dataframe of multitaper power analysis results}

\item{windowParams}{(numeric vector): c(window size (seconds), step size (seconds)) (default: 5 1)}

\item{analysisWindows}{A List of analysis window to generate corresponding heatmap for band mean power analysis}

\item{seizureCondition}{(char) seizure condition name.}
}
\value{
A list of mean power analysis heatmap
}
\description{
compute the mean power analysis over the frequency band using the multitaper results for a condition
for a list of analysis windows
}
\examples{

#Analysis Condition selector
seizureCondition="sz1 (2)"
useBaseline=TRUE
baselineCondition="30s before sz1"
baselineDuration<-10
#Analysis option
#Group1
analysisWindows<- vector("list",2)
analysisWindows[[1]]$frequencyRange<-c(13,30)
analysisWindows[[1]]$timeWindow<-c(0,10)
analysisWindows[[2]]$frequencyRange<-c(30,90)
analysisWindows[[2]]$timeWindow<-c(0,10)
MeanPowBandHeatmaps<-generateHeatmapMeanPowBand(epochList, multitaperResults, analysisWindows, seizureCondition)
}
