% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{generateMultitaper}
\alias{generateMultitaper}
\title{compute the power analysis over the frequency band using the multitaper method of an epoch list}
\usage{
generateMultitaper(
  epochList,
  rangeBand,
  timeBandwidth,
  numTapers,
  windowParams,
  minNfft,
  weighting,
  detrendOpt,
  parallel,
  numWorkers
)
}
\arguments{
\item{epochList}{list of Matrix or Epoch object. iEEG data matrix or Epoch object. If matrix, the row names are the electrode names and the column names are the time points}

\item{rangeBand}{Numeric. Frequency band of the multitaper power analysis}

\item{timeBandwidth}{(numeric): time-half bandwidth product (window duration\emph{half bandwidth of main lobe)
(default: 5 Hz}s)}

\item{numTapers}{(numeric): number of DPSS tapers to use (default: NULL will be computed
as floor(2*time_bandwidth - 1))}

\item{windowParams}{(numeric vector): c(window size (seconds), step size (seconds)) (default: 5 1)}

\item{minNfft}{(numeric) minimum allowable NFFT size, adds zero padding for interpolation (closest 2^x) (default: 0)}

\item{weighting}{(char) weighting of tapers ('unity' (default), 'eigen', 'adapt')}

\item{detrendOpt}{(char): detrend data window ('linear' (default), 'constant', 'off')}

\item{parallel}{(logical): use parallel processing to speed up calculation (default: FALSE). Note: speedup is faster on}

\item{numWorkers}{(logical): number of workers}
}
\value{
A dataframe of power analysis
}
\description{
compute the power analysis over the frequency band using the multitaper method of an epoch list
}
\examples{
# access package Data in OSF
dl<-EpochDownloader(progress = FALSE)
names(dl)
# data download one patient multiple seizure
pt01sz1 <- dl$FragilityData_subpt01_1
pt01sz1
pt01sz2 <- dl$FragilityData_subpt01_2
pt01sz2

# crop Baseline and signal to analyze
useBaseline=TRUE
pt01sz1Baseline<-crop(pt01sz1, start=-30, end=-20)
pt01sz1Cropped<-crop(pt01sz1, start=0, end=20)
pt01sz2Baseline<-crop(pt01sz2, start=-30, end=-20)
pt01sz2Cropped<-crop(pt01sz2, start=0, end=20)
conditions<-c("30s before sz1","sz1","30s before sz2","sz2")
#To visualize patient 01 seizure 1:
pt01sozName <- rownames(pt01sz1Cropped)[rowData(pt01sz1Cropped)$soz]
pt01Display <- c(pt01sozName, "MLT1", "MLT2", "MLT3", "MLT4")

pt01sz1Reordered <- pt01sz1Cropped[pt01Display, ]
pt01sz1BSReordered <- pt01sz1Baseline[pt01Display, ]

pt01sz2Reordered <- pt01sz2Cropped[pt01Display, ]
pt01sz2BSReordered <- pt01sz2Baseline[pt01Display, ]

# create Epoch list to follow RAVE/FREEZ framework (have all frequency range, all seizure results ready in UI background)
epochList<-list(pt01sz1BSReordered,pt01sz1Reordered,pt01sz2BSReordered,pt01sz2Reordered)

names(epochList)<-conditions
}
