% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{meanPowBaselineBand}
\alias{meanPowBaselineBand}
\title{Title}
\usage{
meanPowBaselineBand(
  epoch,
  fs,
  windowParams,
  rangeBand,
  powTimeWindow,
  baseTimeWindow
)
}
\arguments{
\item{epoch}{}

\item{fs}{}

\item{windowParams}{}

\item{rangeBand}{}

\item{powTimeWindow}{}

\item{baseTimeWindow}{}
}
\description{
Title
}
