% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plotERHeatmap}
\alias{plotERHeatmap}
\title{Visualization functions ( ER matrix)}
\usage{
plotERHeatmap(ER, sozIndex = NULL)
}
\arguments{
\item{sozIndex}{Integer or string. A group of electrodes to mark as in the Seizure Onset Zone (SOZ)}

\item{ermaster}{ermaster object from}
}
\value{
A ggplot object
}
\description{
\code{plotERHeatmap}: plot epileptogenic ratio heatmaps with electrodes marked as soz colored
}
\examples{

data("pt01EcoG")

## sozIndex is the index of the electrodes we assume are in the SOZ
sozIndex <- attr(pt01EcoG, "sozIndex")

## precomputed fragility object
data("pt01Frag")

## plot the fragility heatmap
plotFragHeatmap(frag = pt01Frag, sozIndex = sozIndex)

}
