% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powStat.r
\name{powStat}
\alias{powStat}
\title{Compute quantiles, mean and standard deviation for two electrodes group marked as soz non marked as soz}
\usage{
powStat(pow, sozIndex)
}
\arguments{
\item{pow}{Matrix or mean power object. Either a matrix with row as Electrode names and Column as power index,
or a MeanPowBand object from \code{meanPowBaselineBand}}

\item{sozIndex}{Integer.  Vector soz electrodes (for good electrodes)}
}
\value{
list of 5 items with quantile matrix, mean and sdv from both electrodes groups
}
\description{
Compute quantiles, mean and standard deviation for two electrodes group marked as soz non marked as soz
}
\examples{
data("pt01betaBandPow")
data("pt01EcoG")
sozIndex <- attr(pt01EcoG, "sozIndex")
pt01powstat <- powStat(pow = pt01betaBandPow, sozIndex = sozIndex)
}
