% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{meanPowBaselineBand}
\alias{meanPowBaselineBand}
\title{compute the mean power analysis over the frequency band using the multitaper method}
\usage{
meanPowBaselineBand(
  epoch,
  fs = 1000,
  windowParams,
  rangeBand,
  powTimeWindow = c(0, 20),
  baseTimeWindow = c(-30, -20)
)
}
\arguments{
\item{epoch}{Matrix or Epoch object. iEEG data matrix or Epoch object. If matrix, the row names are the electrode names and the column names are the time points}

\item{fs}{Numeric. frequency of signal iEEG acquisition}

\item{windowParams}{Numeric. Window parameters of the multitaper method}

\item{rangeBand}{Numeric. Frequency band of the multitaper power analysis}

\item{powTimeWindow}{Numeric. Time window around seizure onset of the power analysis}

\item{baseTimeWindow}{Numeric. Time window around seizure onset to subtract the mean baselien signal}
}
\value{
A power analysis matrix
}
\description{
compute the mean power analysis over the frequency band using the multitaper method
}
\examples{
# access package Data in OSF
dl<-EpochDownloader(progress = FALSE)
names(dl)
# test EZMultitaper with Epoch package
# data download one patient one seizure
pt01sz1 <- dl$FragilityData_subpt01_1
pt01sz1
# crop
pt01sz1Cropped<-crop(pt01sz1, start=-30, end=20)
# select small number of channel
pt01sozName <- rownames(pt01sz1Cropped)[rowData(pt01sz1Cropped)$soz]
pt01Display <- c(pt01sozName, "MLT1", "MLT2", "MLT3", "MLT4")
pt01sz1Reordered <- pt01sz1Cropped[pt01Display, ]
betaBand<-c(13,30)
rangeBand<-betaBand
powTimeWindow<-c(0,20)
baseTimeWindow<-c(-30,-20)
meta<-metaData(pt01sz1Reordered)
fs<-metaData(pt01sz1Reordered)$samplingRate
windowParams<-c(0.25,0.1)
# compute the mean power analysis over the frequency band (rangeBand) over time window (powTimeWindow) and baselined time window (baseTimeWindow)
betaBandPow<-meanPowBaselineBand( epoch=pt01sz1Reordered, fs=fs, windowParams=windowParams, rangeBand=betaBand, powTimeWindow=powTimeWindow, baseTimeWindow=baseTimeWindow)
}
