% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plotPowHeatmap}
\alias{plotPowHeatmap}
\title{Visualization functions ( Pow matrix)}
\usage{
plotPowHeatmap(pow, sozIndex = NULL)
}
\arguments{
\item{sozIndex}{Integer or string. A group of electrodes to mark as in the Seizure Onset Zone (SOZ)}

\item{Pow}{MeanPowBand object from \code{meanPowBaselineBand}}
}
\value{
A ggplot object
}
\description{
\code{plotPowHeatmap}: plot power band heatmap with electrodes marked as soz colored
}
\examples{

data("pt01EcoG")

## sozIndex is the index of the electrodes we assume are in the SOZ
sozIndex <- attr(pt01EcoG, "sozIndex")

}
