% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plotPowHeatmap}
\alias{plotPowHeatmap}
\alias{plotPowQuantile}
\alias{plotPowDistribution}
\title{Visualization functions ( Pow matrix)}
\usage{
plotPowHeatmap(pow, sozIndex = NULL)

plotPowQuantile(pow, sozIndex = NULL)

plotPowDistribution(pow, sozIndex = NULL)
}
\arguments{
\item{sozIndex}{Integer or string. A group of electrodes to mark as in the Seizure Onset Zone (SOZ)}

\item{Pow}{MeanPowBand object from \code{meanPowBaselineBand}}
}
\value{
A ggplot object
}
\description{
\code{plotPowHeatmap}: plot power band heatmap with electrodes marked as soz colored

\code{plotPowQuantile}: Plot mean power time quantiles for two electrodes group marked as SOZ and reference

\code{plotPowDistribution}: Plot mean power time distribution for two electrodes group marked as SOZ and reference
}
\examples{

data("pt01EcoG")

## sozIndex is the index of the electrodes we assume are in the SOZ
sozIndex <- attr(pt01EcoG, "sozIndex")
## precomputed meanPowBand object
data("pt01betaBandPow")

## plot the mean power heatmap
plotPowBand<-plotPowHeatmap(pow=betaBandPow,sozIndex=sozIndex)
plotPowBand
## plot the mean power quantiles
plotbetaQuantile<-plotPowQuantile(pow = pt01betaBandPow, sozIndex = sozIndex)
plotbetaQuantile
## plot the mean power distribution
plotBetaDistr<-plotPowDistribution(pow = pt01betaBandPow, sozIndex = sozIndex)
plotBetaDistr
}
